backend/requirement.txt
aioredis==1.3.1
apturl==0.5.2
argcomplete==1.8.1
argh==0.26.2
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
black==22.1.0
blessings==1.6
blinker==1.4
bpython==0.18
Brlapi==0.7.0
bs4==0.0.1
certifi==2019.11.28
cffi==1.15.0
channels==3.0.5
channels-redis==3.4.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==37.0.2
cupshelpers==1.0
curtsies==0.3.1
daphne==3.0.2
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.6
django-cors-headers==3.13.0
django-extensions==3.2.0
django-filter==22.1
djangorestframework==3.13.1
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
filelock==3.6.0
gpg===1.13.1-unknown
greenlet==0.4.15
gunicorn==20.0.4
h11==0.13.0
hiredis==2.0.0
HTMLParser==0.0.2
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
ipdb==0.13.9
ipython==8.0.1
jedi==0.18.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
matplotlib-inline==0.1.3
msgpack==1.0.4
mypy-extensions==0.4.3
netifaces==0.10.4
numpy==1.23.0
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.10
outcome==1.2.0
packageshail==0.4
pandas==1.4.3
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==2.4.1
prompt-toolkit==3.0.27
protobuf==3.6.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==4.3.0
service-identity==21.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
stack-data==0.1.4
systemd-python==234
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.0.1
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
urlopen==1.0.0
virtualenv==20.13.3
wadllib==1.3.3
watchdog==0.9.0
wcwidth==0.2.5
webdriver-manager==3.7.1
wsproto==1.1.0
xkit==0.0.0
XlsxWriter==3.0.3
zope.interface==5.4.0
